stylo: Move font-size stuff to values::*::font
